Only show chart once container has been measured
