+++ src/components/LineChart/Chart.scss
@@ -1,0 +1,6 @@
+.Container {
+  position: relative;
+  overflow: hidden;
+  height: 100%;
+  width: 100%;
+}
