--- src/components/LineChart/LineChart.scss
@@ -1,6 +1,0 @@
-@import 'node_modules/@shopify/polaris-tokens/dist/index';
-
-.Container {
-  position: relative;
-  overflow: hidden;
-}
